Restructure menu buttons
