+++ src/components/MenuButton/index.js
@@ -1,0 +1,7 @@
+import DropDownMenuButton from './DropDownMenuButton';
+import OwnSettingsMenuButton from './OwnSettingsMenuButton';
+
+export {
+  DropDownMenuButton,
+  OwnSettingsMenuButton,
+};
